% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuck-you.R
\name{check_total_sum}
\alias{check_total_sum}
\title{Check if total sum matches the hours}
\usage{
check_total_sum(df, total_hours)
}
\arguments{
\item{df}{Data.frame to check}

\item{total_hours}{Total number of hours that should be allocated}
}
\description{
Check if total sum matches the hours
}
