% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuck-you.R
\name{get_timesheet}
\alias{get_timesheet}
\title{Generate random timesheet of work hours spend on specified project activities}
\usage{
get_timesheet(activities, n_hours, specified_activities = NULL)
}
\arguments{
\item{activities}{Vector with activity names}

\item{n_hours}{Total number of hours spent on activities}

\item{specified_activities}{data.frame with columns activity and hours}
}
\description{
Generate random timesheet of work hours spend on specified project activities
}
