% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuck-you.R
\docType{data}
\name{SYRI_ACTIVITIES}
\alias{SYRI_ACTIVITIES}
\title{SYRI list of activities}
\format{
An object of class \code{character} of length 15.
}
\usage{
SYRI_ACTIVITIES
}
\description{
SYRI list of activities
}
\keyword{datasets}
